refactor(interfaces): cleaned up interface namespaces
they were alternating with and without Interface name in the end
