+++ src/Sepes.Infrastructure/Service/Interface/ILookupService.cs
@@ -1,7 +1,7 @@
 ﻿using Sepes.Infrastructure.Dto;
 using System.Collections.Generic;
 
-namespace Sepes.Infrastructure.Service
+namespace Sepes.Infrastructure.Service.Interface
 {
     public interface ILookupService
     {
