+++ src/Sepes.Infrastructure/Service/Interface/IProvisioningQueueService.cs
@@ -1,9 +1,9 @@
 ﻿using Sepes.Infrastructure.Dto;
 using System.Threading.Tasks;
 
-namespace Sepes.Infrastructure.Service
+namespace Sepes.Infrastructure.Service.Interface
 {
-    public interface IResourceProvisioningQueueService
+    public interface IProvisioningQueueService
     {
         Task SendMessageAsync(ProvisioningQueueParentDto message);
