+++ src/Sepes.Infrastructure/Service/Interface/IVariableService.cs
@@ -1,7 +1,7 @@
 ﻿using Sepes.Infrastructure.Model;
 using System.Threading.Tasks;
 
-namespace Sepes.Infrastructure.Service
+namespace Sepes.Infrastructure.Service.Interface
 {
     public interface IVariableService
     {
