+++ src/Sepes.RestApi/Startup.cs
@@ -13,6 +13,7 @@ using Sepes.Infrastructure.Model.Automapper;
 using Sepes.Infrastructure.Model.Config;
 using Sepes.Infrastructure.Model.Context;
 using Sepes.Infrastructure.Service;
+using Sepes.Infrastructure.Service.Azure.Interface;
 using Sepes.Infrastructure.Service.Interface;
 using Sepes.RestApi.Middelware;
 using Sepes.RestApi.Services;
@@ -136,11 +137,11 @@ namespace Sepes.RestApi
             services.AddTransient<IAzureVNetService, AzureVNetService>();
             services.AddTransient<IAzureVMService, AzureVMService>();
             services.AddTransient<IAzureQueueService, AzureQueueService>();
-            services.AddTransient<IResourceProvisioningQueueService, ResourceProvisioningQueueService>();
+            services.AddTransient<IProvisioningQueueService, ResourceProvisioningQueueService>();
             services.AddTransient<IAzureStorageAccountService, AzureStorageAccountService>();
             services.AddTransient<ISandboxResourceProvisioningService, SandboxResourceProvisioningService>();
             services.AddTransient<ISandboxResourceOperationService, SandboxResourceOperationService>();
-            services.AddTransient<IAzureResourceMonitoringService, AzureResourceMonitoringService>();   
+            services.AddTransient<ISandboxResourceMonitoringService, SandboxResourceMonitoringService>();   
             services.AddTransient<IAzureADUsersService, AzureADUsersService>();
             services.AddTransient<IGraphServiceProvider, GraphServiceProvider>();
             services.AddTransient<SandboxResourceOperationService>();
